feature(#98): created email router module and updated app
